+++ codegen/src/rust.rs
@@ -901,7 +901,7 @@ impl<'a> RustGenerator<'a> {
             genln!(self);
         }
 
-        genln!(self, "    pub async fn subscribe_all(&mut self) -> Result<(), aldrin::Error> {{");
+        genln!(self, "    pub async fn subscribe_all(&self) -> Result<(), aldrin::Error> {{");
         for item in svc.items() {
             let ev = match item {
                 ast::ServiceItem::Event(ev) => ev,
@@ -914,7 +914,7 @@ impl<'a> RustGenerator<'a> {
         genln!(self, "    }}");
         genln!(self);
 
-        genln!(self, "    pub async fn unsubscribe_all(&mut self) -> Result<(), aldrin::Error> {{");
+        genln!(self, "    pub async fn unsubscribe_all(&self) -> Result<(), aldrin::Error> {{");
         for item in svc.items() {
             let ev = match item {
                 ast::ServiceItem::Event(ev) => ev,
@@ -936,12 +936,12 @@ impl<'a> RustGenerator<'a> {
             let ev_name = ev.name().value();
             let id = ev.id().value();
 
-            genln!(self, "    pub async fn {}(&mut self) -> Result<(), aldrin::Error> {{", subscribe_event(ev_name));
+            genln!(self, "    pub async fn {}(&self) -> Result<(), aldrin::Error> {{", subscribe_event(ev_name));
             genln!(self, "        self.inner.subscribe({id}).await");
             genln!(self, "    }}");
             genln!(self);
 
-            genln!(self, "    pub async fn {}(&mut self) -> Result<(), aldrin::Error> {{", unsubscribe_event(ev_name));
+            genln!(self, "    pub async fn {}(&self) -> Result<(), aldrin::Error> {{", unsubscribe_event(ev_name));
             genln!(self, "        self.inner.unsubscribe({id}).await");
             genln!(self, "    }}");
             genln!(self);