+++ codegen/src/rust.rs
@@ -1090,7 +1090,7 @@ impl<'a> RustGenerator<'a> {
         genln!(self, "        Ok({} {{ inner }})", svc_name);
         genln!(self, "    }}");
         genln!(self);
-        genln!(self, "    pub fn into_low_level(self) -> aldrin::low_level::Service {{");
+        genln!(self, "    pub fn into_inner(self) -> aldrin::low_level::Service {{");
         genln!(self, "        self.inner");
         genln!(self, "    }}");
         genln!(self);