+++ codegen/CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Add `introspection` flag to `Options`.
 - Add `introspection_if` to `RustOptions`.
 - Support generating introspection.
+- Add `type_id()` to proxies.
 
 ## [0.6.0] - 2024-06-07