+++ codegen/CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+
+### Fixed
+
+- Fix generated code when introspection is enabled unconditionally, that is when `introspection` is
+  `true` but `introspection_if` is `None`.
+
 ## [0.8.0] - 2024-09-22
 
 ### Added