+++ codegen/CHANGELOG.md
@@ -5,7 +5,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## Unreleased
+## [0.8.2] - 2024-10-30
 
 ### Fixed
 
@@ -94,6 +94,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Initial release.
 
+[0.8.2]: https://github.com/dennis-hamester/aldrin/releases/tag/aldrin-codegen-0.8.2
 [0.8.1]: https://github.com/dennis-hamester/aldrin/releases/tag/aldrin-codegen-0.8.1
 [0.8.0]: https://github.com/dennis-hamester/aldrin/releases/tag/aldrin-codegen-0.8.0
 [0.7.0]: https://github.com/dennis-hamester/aldrin/releases/tag/aldrin-codegen-0.7.0